Add locking for cron job that sync to the GitHub repository mirror
Occasionally, the synchronization from the Mercurial repository of Octave on Savannah to the mirror on GitHub stalls for some branches. In that case, the content of the affected branch resembles one of the other branches.

That might be because pulling from Savannah can take very long when it is under DDOS attacks. The synchronization cron job executes every five minutes. If pulling from the three branches ("stable", "default", and "bytecode-interpreter") takes longer than five minutes, multiple processes that try to switch branches could run concurrently.

Trying to avoid the script that does the synchronization was change to create a directory while the script runs. Creating a directory should be an atomic operation on Linux.

Document that change here.